refix mock trainer
