+++ prompt2model/trainer/base.py
@@ -5,11 +5,22 @@ from typing import Any
 
 import datasets
 import transformers
+from transformers import AutoModel, AutoTokenizer
 
 
 # pylint: disable=too-few-public-methods
 class Trainer(ABC):
     """Train a model with a fixed set of hyperparameters."""
+
+    def __init__(self, pretrained_model_name: str):
+        """Initialize a model trainer.
+
+        Args:
+            pretrained_model_name: A HuggingFace model name to use for training.
+        """
+        self.model = AutoModel.from_pretrained(pretrained_model_name)
+        self.tokenizer = AutoTokenizer.from_pretrained(pretrained_model_name)
+        self.wandb = None
 
     @abstractmethod
     def train_model(
